Move CMPSTOList into seperate file
