+++ src/Makefile.omnicore.include
@@ -5,6 +5,7 @@ OMNICORE_H = \
   omnicore/createpayload.h \
   omnicore/createtx.h \
   omnicore/dbbase.h \
+  omnicore/dbstolist.h \
   omnicore/dbtransaction.h \
   omnicore/dex.h \
   omnicore/encoding.h \
@@ -46,6 +47,7 @@ OMNICORE_CPP = \
   omnicore/createpayload.cpp \
   omnicore/createtx.cpp \
   omnicore/dbbase.cpp \
+  omnicore/dbstolist.cpp \
   omnicore/dbtransaction.cpp \
   omnicore/dex.cpp \
   omnicore/encoding.cpp \
